pcq: refactor into library and cli function
This prepares for other programs that use libpcq; I have an idea for
a yaml-driven test fixture...
