+++ src/pcq_lib.c
@@ -1,3 +1,7 @@
+/* SPDX-License-Identifier: Apache-2.0 */
+/*
+ * Copyright (C) 2024 Micron Technology, Inc.  All rights reserved.
+ */
 
 #include <stdio.h>
 #include <sys/types.h>
@@ -33,68 +37,9 @@
 #include "mu_mem.h"
 #include "random_buffer.h"
 #include "famfs.h"
+#include "pcq.h"
 
 extern int mock_flush;
-
-#define PCQ_MAGIC 0xBEEBEE3
-#define PCQ_CONSUMER_MAGIC 0xBEEBEE4
-
-/**
- * struct @pcq
- *
- * @pcq_magic
- * @nbuckets            - number of buckets
- * @bucket_size         - bucket size, inclusive of crc in the last 32 bits
- * @bucket_array_offset - offset within this file of the first bucket
- * @producer_index      - index of the last valid entry; empty if == consumer_index
- * @next_seq            - next seq number (not in same cacche line as producer_index)
- */
-struct pcq {
-	u64 pcq_magic;
-	u64 nbuckets;
-	u64 bucket_size;
-	u64 bucket_array_offset;
-	u64 producer_index;
-	char pad[1024];
-	u64 next_seq;
-	u64 pcq_size;
-};
-
-/**
- * struct @pcq_consumer
- *
- * @pcq_consumer_magic
- * @pad
- * @consumer_index - Consumer index for the queue
- * @pad2
- * @next_seq       - Sequence number for next entry from the queue
- */
-struct pcq_consumer {
-	u32 pcq_consumer_magic;
-	u32 pad;
-	u64 consumer_index;
-	char pad2[1048576];
-	u64 next_seq;
-	u64 pcqc_size;
-};
-
-struct pcq_handle {
-	struct pcq *pcq;
-	struct pcq_consumer *pcqc;
-};
-
-static inline int64_t
-pcq_payload_size(struct pcq *pcq)
-{
-	assert(pcq->pcq_magic == PCQ_MAGIC);
-	return pcq->bucket_size - sizeof(unsigned long) - sizeof(u64);
-}
-
-static inline u64
-pcq_crc_offset(struct pcq *pcq)
-{
-	return pcq->bucket_size - sizeof(unsigned long);
-}
 
 bool
 pcq_valid(struct pcq_handle *pcqh, int verbose)
@@ -155,7 +100,6 @@ pcq_nmessages(struct pcq_handle *pcqh)
  *
  * Caller must free the returned string
  */
-
 static char *
 pcq_consumer_fname(const char *basename)
 {
@@ -294,39 +238,6 @@ out:
 	free(consumer_fname);
 	return 0;
 }
-
-enum pcq_role {
-	PRODUCER,
-	CONSUMER,
-	READONLY,
-};
-
-enum stop_mode {
-	EMPTY, /* consumer only */
-	NMESSAGES,
-	STOP_FLAG,
-};
-
-struct pcq_thread_arg {
-	enum pcq_role role;
-	int verbose;
-	enum stop_mode stop_mode;
-	u64 nmessages;
-	u64 runtime;
-	u64 seed;
-	bool wait;
-	char *basename;
-	int stop_now;
-
-	/* Outputs */
-	u64 nsent;
-	u64 nreceived;
-	u64 nerrors;
-	u64 nfull;  /* # of times full (producer) */
-	u64 nempty; /* # of times empty (consumer) */
-	u64 retries;
-	int result;
-};
 
 
 struct pcq_handle *
@@ -737,14 +648,6 @@ pcq_worker(void *arg)
 	return rc;
 }
 
-struct pcq_status_thread_arg {
-	struct pcq_thread_arg *p; /* producer */
-	struct pcq_thread_arg *c; /* consumer */
-	char *basename;
-	u64 interval;
-	int stop_now;
-};
-
 void *status_worker(void *arg)
 {
 	struct pcq_status_thread_arg *a = arg;
@@ -806,14 +709,6 @@ out:
 	return rc;
 }
 
-enum pcq_perm {
-	pcq_perm_nop=0,
-	pcq_perm_none,
-	pcq_perm_both,
-	pcq_perm_producer,
-	pcq_perm_consumer,
-};
-
 int
 pcq_set_perm(const char *filename, enum pcq_perm role)
 {
@@ -858,445 +753,4 @@ pcq_set_perm(const char *filename, enum pcq_perm role)
 out:
 	free(consumer_fname);
 	return rc;
-}
-
-/* XXX Move this to famfs_lib_util.c or some such - it is now shared with the cli */
-static s64 get_multiplier(const char *endptr)
-{
-	size_t multiplier = 1;
-
-	if (!endptr)
-		return 1;
-
-	switch (*endptr) {
-	case 'k':
-	case 'K':
-		multiplier = 1024;
-		break;
-	case 'm':
-	case 'M':
-		multiplier = 1024 * 1024;
-		break;
-	case 'g':
-	case 'G':
-		multiplier = 1024 * 1024 * 1024;
-		break;
-	case 0:
-		return 1;
-	}
-	++endptr;
-	if (*endptr) /* If the unit was not the last char in string, it's an error */
-		return -1;
-	return multiplier;
-}
-
-void
-pcq_usage(int   argc,
-	    char *argv[])
-{
-	char *progname = argv[0];
-
-	printf("\n"
-	       "pcq: Run a producer/consumer queue test\n"
-	       "\n"
-	       "This test creates a set of files to use as a producer/consumer queue\n"
-	       "and sends messages through the queue. You can run one copy of this program\n"
-	       "that does both the producer & consumer functions, but to test shared memory\n"
-	       "you need to run one copy of this program as producer an a second copy (on a\n"
-	       "different node) as consumer.\n"
-	       "\n"
-	       "A producer/consumer queue is implemented as a set of files. The producer file\n"
-	       "contains the producer index and the buckets. The consumer file contains the\n"
-	       "consumer index. The producer maps the producer file writable and the consumer\n"
-	       "file read-only; the consumer does the opposite.\n"
-	       "\n"
-	       "EXAMPLES:\n"
-	       "Create a producer/consumer queue with 4096 buckets which are 1K each:\n"
-	       "    %s --create --bsize 1024 --nbuckets 4K <queuename>\n"
-	       "\n"
-	       "Just run a producer:\n"
-	       "    %s --producer [Args] /mnt/famfs/<queuename>\n"
-	       "\n"
-	       "Run a consumer:\n"
-	       "    %s --consumer [Args] /mnt/famfs/<queuename>\n"
-	       "\n"
-	       "Run a producer and a consumer from a single process:\n"
-	       "    %s --producer --consumer [Args] /mnt/famfs/<queuename>\n"
-	       "\n"
-	       "Drain a pcq\n"
-	       "    %s --drain [Args] /mnt/famfs/<queuename>\n"
-	       "\n"
-	       "Check the state of a producer/consumer queue (maps both fies read-only:\n"
-	       "    %s --info [Args] /mnt/famfs/<queuename>\n"
-	       "\n"
-	       "Arguments:\n"
-	       "\n"
-	       "Queue Creation:\n"
-	       "    -C|--create               - Create a producer/consumer queue\n"
-	       "    -b|--bsize <bucketsize>   - size of messages including sequence number\n"
-	       "                                and crc (ignored if queue already exists)\n"
-	       "    -n|--nbuckets <nnbuckets> - Number of buckets in the queue\n"
-	       "                                (ignored if queue already exists)\n"
-	       "\n"
-	       "Queue permissions:\n"
-	       "    -P|--setperm <p|c|b|n>    - Set permissions on a queue for (p)roducer or\n"
-	       "                                (c)onsumer, (b)oth or (n)either on this node.\n"
-	       "                                Must run  separately from create|producer|consumer|drain\n"
-	       "\n"
-	       "Running producers and consumers:\n"
-	       "    -N|--nmessages <n>        - Number of messages to send and/or receive\n"
-	       "    -t|--time <seconds>       - Run for the specified duration\n"
-	       "    -S|--seed <seed>          - Use seed to generate payload\n"
-	       "    -p|--producer             - Run the producer\n"
-	       "    -c|--consumer             - Run the consumer\n"
-	       "    -s|--status <interval>    - Print status at the specified interval\n"
-	       "\n"
-	       "Special options:\n"
-	       "    -i|--info                 - Dump the state of a queue\n"
-	       "    -d|--drain                - Run a consumer to drain a queue to empty and\n"
-	       "                                then exit. (Note this probably won't do what\n"
-	       "                                you want if a producer is running...)\n"
-	       "    -D|--dontflush            - Don't issue processor cache flushes and\n"
-	       "                                invalidates\n"
-	       "    -f|--statusfile           - Write exit status to file (for testing)\n"
-	       "    -?                        - Print this message\n"
-	       "\n", progname, progname, progname, progname, progname, progname);
-}
-
-int
-main(int argc, char **argv)
-{
-	pthread_t producer_thread, consumer_thread, status_thread;
-	struct pcq_status_thread_arg status = { 0 };
-	struct pcq_thread_arg prod = { 0 };
-	struct pcq_thread_arg cons = { 0 };
-	enum pcq_perm role = pcq_perm_nop;
-	char *statusfname = NULL;
-	FILE *statusfile = NULL;
-	u64 status_interval = 0;
-	char *filename = NULL;
-	bool producer = false;
-	bool consumer = false;
-	bool create = false;
-	u64 bucket_size = 0;
-	bool drain = false;
-	u64 nmessages = 0;
-	bool info = false;
-	u64 nbuckets = 0;
-	int wait = true;
-	int runtime = 0;
-	int verbose = 0;
-	s64 seed = 0;
-	int arg_ct;
-	int c, rc;
-	s64 mult;
-
-	struct option pcq_options[] = {
-		/* These options set a flag. */
-		{"bsize",       required_argument,        0,  'b'},
-		{"nbuckets",    required_argument,        0,  'n'},
-		{"seed",        required_argument,        0,  'S'},
-		{"nmessages",   required_argument,        0,  'N'},
-		{"statusfile",  required_argument,        0,  'f'},
-		{"time",        required_argument,        0,  't'},
-		{"status",      required_argument,        0,  's'},
-		{"setperm",     required_argument,        0,  'P'},
-
-		{"create",      no_argument,              0,  'C'},
-		{"producer",    no_argument,              0,  'p'},
-		{"consumer",    no_argument,              0,  'c'},
-		{"info",        no_argument,              0,  'i'},
-		{"drain",       no_argument,              0,  'd'},
-		{"dontflush",   no_argument,              0,  'D'},
-		/* These options don't set a flag.
-		 * We distinguish them by their indices.
-		 */
-		/*{"dryrun",       no_argument,       0, 'n'}, */
-		{0, 0, 0, 0}
-	};
-
-	/* Note: the "+" at the beginning of the arg string tells getopt_long
-	 * to return -1 when it sees something that is not recognized option
-	 * (e.g. the command that will mux us off to the command handlers
-	 */
-	while ((c = getopt_long(argc, argv, "+b:s:S:n:N:f:t:s:CdpcwDih?v",
-				pcq_options, &optind)) != EOF) {
-		char *endptr;
-
-		arg_ct++;
-		switch (c) {
-
-		case 'i':
-			info = true;
-			break;
-
-			/* Create operations */
-		case 'C':
-			create = true;
-			break;
-
-		case 'b':
-			bucket_size = strtoull(optarg, &endptr, 0);
-			mult = get_multiplier(endptr);
-			if (mult > 0)
-				bucket_size *= mult;
-			printf("bucket_size=%lld\n", bucket_size);
-			break;
-
-		case 'n':
-			nbuckets = strtoull(optarg, &endptr, 0);
-			mult = get_multiplier(endptr);
-			if (mult > 0)
-				nbuckets *= mult;
-			printf("nbuckets=%lld\n", nbuckets);
-			break;
-
-		case 'S':
-			seed = strtoull(optarg, 0, 0);
-			break;
-
-		case 's':
-			status_interval = strtoull(optarg, 0, 0);
-			break;
-
-		case 't':
-			runtime = strtoull(optarg, 0, 0);
-			break;
-
-		case 'p':
-			producer = true;
-			break;
-
-		case 'c':
-			consumer = true;
-			break;
-
-		case 'd':
-			drain = true;
-			wait = false;
-			break;
-
-		case 'v':
-			verbose++;
-			break;
-
-		case 'N':
-			nmessages = strtoull(optarg, &endptr, 0);
-			mult = get_multiplier(endptr);
-			if (mult > 0)
-				nmessages *= mult;
-			break;
-
-		case 'f':
-			/* Write execution status to this file (for testing) */
-			statusfname= optarg;
-			break;
-
-		case 'P':
-			if (strcmp(optarg, "p") == 0)
-				role = pcq_perm_producer;
-			else if (strcmp(optarg, "c") == 0)
-				role = pcq_perm_consumer;
-			else if (strcmp(optarg, "b") == 0)
-				role = pcq_perm_both;
-			else if (strcmp(optarg, "n") == 0)
-				role = pcq_perm_none;
-			else {
-				fprintf(stderr, "%s: invalid --setperm arg (%s)\n",
-					__func__, optarg);
-				pcq_usage(argc, argv);
-				return -1;
-			}
-			break;
-
-		case 'D':
-			mock_flush = 1;
-			break;
-
-		case 'h':
-		case '?':
-			pcq_usage(argc, argv);
-			return 0;
-		}
-	}
-
-	if (info && (create || producer || consumer || drain)) {
-		fprintf(stderr, "%s: info not compatible with operating on a pcq\n\n",
-			argv[0]);
-		pcq_usage(argc, argv);
-		return -1;
-	}
-	if (create && (bucket_size == 0 || nbuckets == 0)) {
-		fprintf(stderr, "%s: create requires bsize and nbuckets\n\n", argv[0]);
-		pcq_usage(argc, argv);
-		return -1;
-	}
-	if (drain && (wait || producer || nmessages || runtime)) {
-		fprintf(stderr,
-			"%s: drain can't be used with producer, time or nmessages options\n\n",
-			argv[0]);
-		pcq_usage(argc, argv);
-		return -1;
-	}
-	if (runtime && nmessages) {
-		fprintf(stderr,
-			"%s: the --nmessages and --time args cannot be used together\n\n",
-			__func__);
-		pcq_usage(argc, argv);
-		return -1;
-	}
-	if (optind > (argc - 1)) {
-		fprintf(stderr, "Must specify base filename\n\n");
-		return -1;
-	}
-	filename = argv[optind++];
-
-	if (statusfname) {
-		truncate(statusfname, 0);
-		statusfile = fopen(statusfname, "w+");
-		assert(statusfile);
-	}
-	if (role != pcq_perm_nop && (create || producer || consumer || drain)) {
-		fprintf(stderr,
-			"--setperm is incompatible with --create|--drain|--producer|--consumer");
-		pcq_usage(argc, argv);
-		return -1;
-	}
-
-	if (role != pcq_perm_nop)
-		return pcq_set_perm(filename, role);
-
-	if (create)
-		return pcq_create(filename, nbuckets, bucket_size, verbose);
-
-	if (info)
-		return get_queue_info(filename, statusfile, verbose);
-	if (drain) {
-		struct pcq_thread_arg ta = { 0 };
-
-		ta.role = CONSUMER;
-		ta.stop_mode = EMPTY;
-		ta.basename = filename;
-		ta.verbose = verbose;
-
-		printf("pcq:    %s\n", filename);
-		rc = run_consumer(&ta);
-		printf("pcq drain: nreceived=%lld nerrors=%lld nempty=%lld retries=%lld\n",
-		       ta.nreceived, ta.nerrors, ta.nempty, ta.retries);
-		if (ta.nerrors) {
-			if (statusfile) {
-				fprintf(statusfile, "%lld", -ta.nerrors);
-				fclose(statusfile);
-			}
-			return -(int)ta.nerrors;
-		}
-
-		if (statusfile) {
-			printf("pcq: drained %lld messages from queue %s, with no errors\n",
-			       ta.nreceived, filename);
-			fprintf(statusfile, "%lld", ta.nreceived);
-			fclose(statusfile);
-		}
-		return rc;
-	}
-
-	/*
-	 * Start the producer thread if needed
-	 */
-	assert(wait);
-	if (producer) {
-		prod.role = PRODUCER;
-		prod.stop_mode = (runtime) ? STOP_FLAG : NMESSAGES;
-		prod.nmessages = nmessages;
-		prod.runtime = runtime;
-		prod.basename = filename;
-		prod.seed = seed;
-		prod.wait = wait;
-		prod.verbose = verbose;
-		rc = pthread_create(&producer_thread, NULL, pcq_worker, (void *)&prod);
-		if (rc) {
-			fprintf(stderr, "%s: failed to start producer thread\n", __func__);
-		}
-	}
-
-	/*
-	 * Start the consumer thread
-	 */
-	if (consumer) {
-		cons.role = CONSUMER;
-		cons.stop_mode = (runtime) ? STOP_FLAG : NMESSAGES;
-		cons.nmessages = nmessages;
-		prod.runtime = runtime;
-		cons.basename = filename;
-		cons.seed = seed;
-		cons.wait = wait;
-		cons.verbose = verbose;
-		rc = pthread_create(&consumer_thread, NULL, pcq_worker, (void *)&cons);
-		if (rc) {
-			fprintf(stderr, "%s: failed to start consumer thread\n", __func__);
-		}
-	}
-
-	if (status_interval) {
-		status.p = &prod;
-		status.c = &cons;
-		status.basename = filename;
-		status.interval = status_interval;
-		status.stop_now = 0;
-		
-		rc = pthread_create(&status_thread, NULL, status_worker, (void *)&status);
-		if (rc) {
-			fprintf(stderr, "%s: failed to start consumer thread\n", __func__);
-		}
-	}
-
-	if (runtime) {
-		sleep(runtime);
-		prod.stop_now = 1;
-		cons.stop_now = 1;
-		status.stop_now = 1;
-	}
-
-	if (producer) {
-		rc = pthread_join(producer_thread, NULL);
-		if (rc)
-			fprintf(stderr, "%s: failed to join producer thread\n", __func__);
-	}
-	if (consumer) {
-		rc = pthread_join(consumer_thread, NULL);
-		if (rc)
-			fprintf(stderr, "%s: failed to join consumer thread\n", __func__);
-	}
-	if (status_interval) {
-		status.stop_now = 1;
-		rc = pthread_join(status_thread, NULL);
-		if (rc)
-			fprintf(stderr, "%s: failed to join consumer thread\n", __func__);
-	}
-
-	printf("pcq:    %s\n", filename);
-	printf("pcq producer: nsent=%lld nerrors=%lld nfull=%lld\n",
-	       prod.nsent, prod.nerrors, prod.nfull);
-	printf("pcq consumer: nreceived=%lld nerrors=%lld nempty=%lld retries=%lld\n",
-	       cons.nreceived, cons.nerrors, cons.nempty, cons.retries);
-
-	if (prod.nerrors || cons.nerrors) {
-		if (statusfile) {
-			/* If there are errors, the statusfile will contain the negative
-			 * sum of the producer and consumer errors
-			 */
-			fprintf(statusfile, "%lld", -(prod.nerrors + cons.nerrors));
-			fclose(statusfile);
-		}
-		return -(int)(prod.nerrors + cons.nerrors);
-	}
-
-	if (statusfile) {
-		/* If there are no errors, the statusfile will contain the sum
-		 * of messages sent and received
-		 */
-		fprintf(statusfile, "%lld", (prod.nsent + cons.nreceived));
-		fclose(statusfile);
-	}
-	return prod.result + cons.result;
 }
